requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

torch==2.2.2+cpu
torchvision==0.17.2+cpu

ultralytics==8.3.0
opencv-python-headless==4.9.0.80
pillow==10.4.0
numpy==1.26.4
psutil==5.9.8
PyYAML==6.0.1
requests==2.32.3
tqdm==4.66.5
fastapi==0.112.2
uvicorn[standard]==0.30.6
